Deleting Directories
The last step of this lesson's Exercise asks to delete a directory, and the Knowledge Check section's link for how to do so brings the student to the Conquering the Command Line chapter. But the chapter only directly teaches how to delete files, where attempting the same command but applied to a directory gives an error message. A link to an article that more directly teaches this difference has been added to the Assignment.